ChebyHash: Split ChebyHash into 32-bit and experimental 64-bit variants

+++ src/ChebyHash/ChebyHash64.hpp
@@ -8,6 +8,6 @@
 // Do not use for real passwords, signatures,
 // authentication, key derivation, or secure storage.
 
-void round(std::uint32_t& x, std::uint32_t& y);
-std::uint32_t mix(std::uint32_t x, std::uint32_t y);
+void round(std::uint64_t& x, std::uint64_t& y);
+std::uint32_t mix(std::uint64_t x, std::uint64_t y);
 void ChebyHashTester();
